Add Url finder node visitor
