+++ misc/ls-extensions/modules/performance-analyzer-services/src/main/java/io/ballerina/ProgramAnalyzerNodeVisitor.java
@@ -271,7 +271,6 @@ public class ProgramAnalyzerNodeVisitor extends NodeVisitor {
     private String file;
     private boolean withinRange = false;
     private Range range;
-    private String currentToken;
 
     public ProgramAnalyzerNodeVisitor() {
 
@@ -373,12 +372,6 @@ public class ProgramAnalyzerNodeVisitor extends NodeVisitor {
 
         this.visitSyntaxNode(functionDefinitionNode);
         withinRange = false;
-    }
-
-    @Override
-    public void visit(ServiceDeclarationNode serviceDeclarationNode) {
-
-        visitSyntaxNode(serviceDeclarationNode);
     }
 
     @Override
@@ -543,975 +536,6 @@ public class ProgramAnalyzerNodeVisitor extends NodeVisitor {
         visitSyntaxNode(assignmentStatementNode);
     }
 
-    @Override
-    public void visit(TypedBindingPatternNode typedBindingPatternNode) {
-
-    }
-
-    @Override
-    public void visit(ImportDeclarationNode importDeclarationNode) {
-
-    }
-
-    @Override
-    public void visit(ListenerDeclarationNode listenerDeclarationNode) {
-
-    }
-
-    @Override
-    public void visit(TypeDefinitionNode typeDefinitionNode) {
-
-    }
-
-    @Override
-    public void visit(CompoundAssignmentStatementNode compoundAssignmentStatementNode) {
-
-    }
-
-    @Override
-    public void visit(BlockStatementNode blockStatementNode) {
-
-        visitSyntaxNode(blockStatementNode);
-    }
-
-    @Override
-    public void visit(BreakStatementNode breakStatementNode) {
-
-    }
-
-    @Override
-    public void visit(FailStatementNode failStatementNode) {
-
-    }
-
-    @Override
-    public void visit(ExpressionStatementNode expressionStatementNode) {
-
-    }
-
-    @Override
-    public void visit(ContinueStatementNode continueStatementNode) {
-
-    }
-
-    @Override
-    public void visit(ExternalFunctionBodyNode externalFunctionBodyNode) {
-
-    }
-
-    @Override
-    public void visit(ElseBlockNode elseBlockNode) {
-
-    }
-
-    @Override
-    public void visit(PanicStatementNode panicStatementNode) {
-
-    }
-
-    @Override
-    public void visit(ReturnStatementNode returnStatementNode) {
-
-    }
-
-    @Override
-    public void visit(LocalTypeDefinitionStatementNode localTypeDefinitionStatementNode) {
-
-    }
-
-    @Override
-    public void visit(LockStatementNode lockStatementNode) {
-
-    }
-
-    @Override
-    public void visit(ForkStatementNode forkStatementNode) {
-
-    }
-
-    @Override
-    public void visit(BinaryExpressionNode binaryExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(BracedExpressionNode bracedExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(CheckExpressionNode checkExpressionNode) {
-
-        visitSyntaxNode(checkExpressionNode);
-    }
-
-    @Override
-    public void visit(FieldAccessExpressionNode fieldAccessExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(FunctionCallExpressionNode functionCallExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(MethodCallExpressionNode methodCallExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(MappingConstructorExpressionNode mappingConstructorExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(IndexedExpressionNode indexedExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(TypeofExpressionNode typeofExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(UnaryExpressionNode unaryExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(ComputedNameFieldNode computedNameFieldNode) {
-
-    }
-
-    @Override
-    public void visit(ConstantDeclarationNode constantDeclarationNode) {
-
-    }
-
-    @Override
-    public void visit(DefaultableParameterNode defaultableParameterNode) {
-
-    }
-
-    @Override
-    public void visit(RequiredParameterNode requiredParameterNode) {
-
-    }
-
-    @Override
-    public void visit(IncludedRecordParameterNode includedRecordParameterNode) {
-
-    }
-
-    @Override
-    public void visit(RestParameterNode restParameterNode) {
-
-    }
-
-    @Override
-    public void visit(ImportOrgNameNode importOrgNameNode) {
-
-    }
-
-    @Override
-    public void visit(ImportPrefixNode importPrefixNode) {
-
-    }
-
-    @Override
-    public void visit(SpecificFieldNode specificFieldNode) {
-
-    }
-
-    @Override
-    public void visit(SpreadFieldNode spreadFieldNode) {
-
-    }
-
-    @Override
-    public void visit(NamedArgumentNode namedArgumentNode) {
-
-        visitSyntaxNode(namedArgumentNode);
-    }
-
-    @Override
-    public void visit(PositionalArgumentNode positionalArgumentNode) {
-
-        visitSyntaxNode(positionalArgumentNode);
-    }
-
-    @Override
-    public void visit(RestArgumentNode restArgumentNode) {
-
-    }
-
-    @Override
-    public void visit(InferredTypedescDefaultNode inferredTypedescDefaultNode) {
-
-    }
-
-    @Override
-    public void visit(ObjectTypeDescriptorNode objectTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(ObjectConstructorExpressionNode objectConstructorExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(RecordTypeDescriptorNode recordTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(ReturnTypeDescriptorNode returnTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(NilTypeDescriptorNode nilTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(OptionalTypeDescriptorNode optionalTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(ObjectFieldNode objectFieldNode) {
-
-    }
-
-    @Override
-    public void visit(RecordFieldNode recordFieldNode) {
-
-    }
-
-    @Override
-    public void visit(RecordFieldWithDefaultValueNode recordFieldWithDefaultValueNode) {
-
-    }
-
-    @Override
-    public void visit(RecordRestDescriptorNode recordRestDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(TypeReferenceNode typeReferenceNode) {
-
-    }
-
-    @Override
-    public void visit(AnnotationNode annotationNode) {
-
-    }
-
-    @Override
-    public void visit(MetadataNode metadataNode) {
-
-    }
-
-    @Override
-    public void visit(TypeTestExpressionNode typeTestExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(MapTypeDescriptorNode mapTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(NilLiteralNode nilLiteralNode) {
-
-    }
-
-    @Override
-    public void visit(AnnotationDeclarationNode annotationDeclarationNode) {
-
-    }
-
-    @Override
-    public void visit(AnnotationAttachPointNode annotationAttachPointNode) {
-
-    }
-
-    @Override
-    public void visit(XMLNamespaceDeclarationNode xMLNamespaceDeclarationNode) {
-
-    }
-
-    @Override
-    public void visit(ModuleXMLNamespaceDeclarationNode moduleXMLNamespaceDeclarationNode) {
-
-    }
-
-    @Override
-    public void visit(FunctionBodyBlockNode functionBodyBlockNode) {
-
-        visitSyntaxNode(functionBodyBlockNode);
-    }
-
-    @Override
-    public void visit(NamedWorkerDeclarationNode namedWorkerDeclarationNode) {
-
-    }
-
-    @Override
-    public void visit(NamedWorkerDeclarator namedWorkerDeclarator) {
-
-    }
-
-    @Override
-    public void visit(BasicLiteralNode basicLiteralNode) {
-
-        visitSyntaxNode(basicLiteralNode);
-    }
-
-    @Override
-    public void visit(SimpleNameReferenceNode simpleNameReferenceNode) {
-
-    }
-
-    @Override
-    public void visit(QualifiedNameReferenceNode qualifiedNameReferenceNode) {
-
-    }
-
-    @Override
-    public void visit(BuiltinSimpleNameReferenceNode builtinSimpleNameReferenceNode) {
-
-    }
-
-    @Override
-    public void visit(TrapExpressionNode trapExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(ListConstructorExpressionNode listConstructorExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(TypeCastExpressionNode typeCastExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(TypeCastParamNode typeCastParamNode) {
-
-    }
-
-    @Override
-    public void visit(UnionTypeDescriptorNode unionTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(TableConstructorExpressionNode tableConstructorExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(KeySpecifierNode keySpecifierNode) {
-
-    }
-
-    @Override
-    public void visit(StreamTypeDescriptorNode streamTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(StreamTypeParamsNode streamTypeParamsNode) {
-
-    }
-
-    @Override
-    public void visit(LetExpressionNode letExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(LetVariableDeclarationNode letVariableDeclarationNode) {
-
-    }
-
-    @Override
-    public void visit(TemplateExpressionNode templateExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(XMLElementNode xMLElementNode) {
-
-    }
-
-    @Override
-    public void visit(XMLStartTagNode xMLStartTagNode) {
-
-    }
-
-    @Override
-    public void visit(XMLEndTagNode xMLEndTagNode) {
-
-    }
-
-    @Override
-    public void visit(XMLSimpleNameNode xMLSimpleNameNode) {
-
-    }
-
-    @Override
-    public void visit(XMLQualifiedNameNode xMLQualifiedNameNode) {
-
-    }
-
-    @Override
-    public void visit(XMLEmptyElementNode xMLEmptyElementNode) {
-
-    }
-
-    @Override
-    public void visit(InterpolationNode interpolationNode) {
-
-    }
-
-    @Override
-    public void visit(XMLTextNode xMLTextNode) {
-
-    }
-
-    @Override
-    public void visit(XMLAttributeNode xMLAttributeNode) {
-
-    }
-
-    @Override
-    public void visit(XMLAttributeValue xMLAttributeValue) {
-
-    }
-
-    @Override
-    public void visit(XMLComment xMLComment) {
-
-    }
-
-    @Override
-    public void visit(XMLProcessingInstruction xMLProcessingInstruction) {
-
-    }
-
-    @Override
-    public void visit(TableTypeDescriptorNode tableTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(TypeParameterNode typeParameterNode) {
-
-    }
-
-    @Override
-    public void visit(KeyTypeConstraintNode keyTypeConstraintNode) {
-
-    }
-
-    @Override
-    public void visit(FunctionTypeDescriptorNode functionTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(FunctionSignatureNode functionSignatureNode) {
-
-    }
-
-    @Override
-    public void visit(ExplicitAnonymousFunctionExpressionNode explicitAnonymousFunctionExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(ExpressionFunctionBodyNode expressionFunctionBodyNode) {
-
-    }
-
-    @Override
-    public void visit(TupleTypeDescriptorNode tupleTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(ParenthesisedTypeDescriptorNode parenthesisedTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(ExplicitNewExpressionNode explicitNewExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(ImplicitNewExpressionNode implicitNewExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(ParenthesizedArgList parenthesizedArgList) {
-
-        visitSyntaxNode(parenthesizedArgList);
-    }
-
-    @Override
-    public void visit(QueryConstructTypeNode queryConstructTypeNode) {
-
-    }
-
-    @Override
-    public void visit(FromClauseNode fromClauseNode) {
-
-    }
-
-    @Override
-    public void visit(WhereClauseNode whereClauseNode) {
-
-    }
-
-    @Override
-    public void visit(LetClauseNode letClauseNode) {
-
-    }
-
-    @Override
-    public void visit(JoinClauseNode joinClauseNode) {
-
-    }
-
-    @Override
-    public void visit(OnClauseNode onClauseNode) {
-
-    }
-
-    @Override
-    public void visit(LimitClauseNode limitClauseNode) {
-
-    }
-
-    @Override
-    public void visit(OnConflictClauseNode onConflictClauseNode) {
-
-    }
-
-    @Override
-    public void visit(QueryPipelineNode queryPipelineNode) {
-
-    }
-
-    @Override
-    public void visit(SelectClauseNode selectClauseNode) {
-
-    }
-
-    @Override
-    public void visit(QueryExpressionNode queryExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(QueryActionNode queryActionNode) {
-
-    }
-
-    @Override
-    public void visit(IntersectionTypeDescriptorNode intersectionTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(ImplicitAnonymousFunctionParameters implicitAnonymousFunctionParameters) {
-
-    }
-
-    @Override
-    public void visit(ImplicitAnonymousFunctionExpressionNode implicitAnonymousFunctionExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(StartActionNode startActionNode) {
-
-    }
-
-    @Override
-    public void visit(FlushActionNode flushActionNode) {
-
-    }
-
-    @Override
-    public void visit(SingletonTypeDescriptorNode singletonTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(MethodDeclarationNode methodDeclarationNode) {
-
-    }
-
-    @Override
-    public void visit(CaptureBindingPatternNode captureBindingPatternNode) {
-
-    }
-
-    @Override
-    public void visit(WildcardBindingPatternNode wildcardBindingPatternNode) {
-
-    }
-
-    @Override
-    public void visit(ListBindingPatternNode listBindingPatternNode) {
-
-    }
-
-    @Override
-    public void visit(MappingBindingPatternNode mappingBindingPatternNode) {
-
-    }
-
-    @Override
-    public void visit(FieldBindingPatternFullNode fieldBindingPatternFullNode) {
-
-    }
-
-    @Override
-    public void visit(FieldBindingPatternVarnameNode fieldBindingPatternVarnameNode) {
-
-    }
-
-    @Override
-    public void visit(RestBindingPatternNode restBindingPatternNode) {
-
-    }
-
-    @Override
-    public void visit(ErrorBindingPatternNode errorBindingPatternNode) {
-
-    }
-
-    @Override
-    public void visit(NamedArgBindingPatternNode namedArgBindingPatternNode) {
-
-    }
-
-    @Override
-    public void visit(AsyncSendActionNode asyncSendActionNode) {
-
-    }
-
-    @Override
-    public void visit(SyncSendActionNode syncSendActionNode) {
-
-    }
-
-    @Override
-    public void visit(ReceiveActionNode receiveActionNode) {
-
-    }
-
-    @Override
-    public void visit(ReceiveFieldsNode receiveFieldsNode) {
-
-    }
-
-    @Override
-    public void visit(RestDescriptorNode restDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(DoubleGTTokenNode doubleGTTokenNode) {
-
-    }
-
-    @Override
-    public void visit(TrippleGTTokenNode trippleGTTokenNode) {
-
-    }
-
-    @Override
-    public void visit(WaitActionNode waitActionNode) {
-
-    }
-
-    @Override
-    public void visit(WaitFieldsListNode waitFieldsListNode) {
-
-    }
-
-    @Override
-    public void visit(WaitFieldNode waitFieldNode) {
-
-    }
-
-    @Override
-    public void visit(AnnotAccessExpressionNode annotAccessExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(OptionalFieldAccessExpressionNode optionalFieldAccessExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(ConditionalExpressionNode conditionalExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(EnumDeclarationNode enumDeclarationNode) {
-
-    }
-
-    @Override
-    public void visit(EnumMemberNode enumMemberNode) {
-
-    }
-
-    @Override
-    public void visit(ArrayTypeDescriptorNode arrayTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(TransactionStatementNode transactionStatementNode) {
-
-    }
-
-    @Override
-    public void visit(RollbackStatementNode rollbackStatementNode) {
-
-    }
-
-    @Override
-    public void visit(RetryStatementNode retryStatementNode) {
-
-    }
-
-    @Override
-    public void visit(CommitActionNode commitActionNode) {
-
-    }
-
-    @Override
-    public void visit(TransactionalExpressionNode transactionalExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(ByteArrayLiteralNode byteArrayLiteralNode) {
-
-    }
-
-    @Override
-    public void visit(XMLFilterExpressionNode xMLFilterExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(XMLStepExpressionNode xMLStepExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(XMLNamePatternChainingNode xMLNamePatternChainingNode) {
-
-    }
-
-    @Override
-    public void visit(XMLAtomicNamePatternNode xMLAtomicNamePatternNode) {
-
-    }
-
-    @Override
-    public void visit(TypeReferenceTypeDescNode typeReferenceTypeDescNode) {
-
-    }
-
-    @Override
-    public void visit(MatchStatementNode matchStatementNode) {
-
-    }
-
-    @Override
-    public void visit(MatchClauseNode matchClauseNode) {
-
-    }
-
-    @Override
-    public void visit(MatchGuardNode matchGuardNode) {
-
-    }
-
-    @Override
-    public void visit(DistinctTypeDescriptorNode distinctTypeDescriptorNode) {
-
-    }
-
-    @Override
-    public void visit(ListMatchPatternNode listMatchPatternNode) {
-
-    }
-
-    @Override
-    public void visit(RestMatchPatternNode restMatchPatternNode) {
-
-    }
-
-    @Override
-    public void visit(MappingMatchPatternNode mappingMatchPatternNode) {
-
-    }
-
-    @Override
-    public void visit(FieldMatchPatternNode fieldMatchPatternNode) {
-
-    }
-
-    @Override
-    public void visit(ErrorMatchPatternNode errorMatchPatternNode) {
-
-    }
-
-    @Override
-    public void visit(NamedArgMatchPatternNode namedArgMatchPatternNode) {
-
-    }
-
-    @Override
-    public void visit(MarkdownDocumentationNode markdownDocumentationNode) {
-
-    }
-
-    @Override
-    public void visit(MarkdownDocumentationLineNode markdownDocumentationLineNode) {
-
-    }
-
-    @Override
-    public void visit(MarkdownParameterDocumentationLineNode markdownParameterDocumentationLineNode) {
-
-    }
-
-    @Override
-    public void visit(BallerinaNameReferenceNode ballerinaNameReferenceNode) {
-
-    }
-
-    @Override
-    public void visit(InlineCodeReferenceNode inlineCodeReferenceNode) {
-
-    }
-
-    @Override
-    public void visit(MarkdownCodeBlockNode markdownCodeBlockNode) {
-
-    }
-
-    @Override
-    public void visit(MarkdownCodeLineNode markdownCodeLineNode) {
-
-    }
-
-    @Override
-    public void visit(OrderByClauseNode orderByClauseNode) {
-
-    }
-
-    @Override
-    public void visit(OrderKeyNode orderKeyNode) {
-
-    }
-
-    @Override
-    public void visit(OnFailClauseNode onFailClauseNode) {
-
-    }
-
-    @Override
-    public void visit(DoStatementNode doStatementNode) {
-
-    }
-
-    @Override
-    public void visit(ClassDefinitionNode classDefinitionNode) {
-
-    }
-
-    @Override
-    public void visit(ResourcePathParameterNode resourcePathParameterNode) {
-
-    }
-
-    @Override
-    public void visit(RequiredExpressionNode requiredExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(ErrorConstructorExpressionNode errorConstructorExpressionNode) {
-
-    }
-
-    @Override
-    public void visit(ParameterizedTypeDescriptorNode parameterizedTypeDescriptorNode) {
-
-    }
-
-    @Override
-    protected void visitSyntaxNode(io.ballerina.compiler.syntax.tree.Node node) {
-
-        super.visitSyntaxNode(node);
-    }
-
-    @Override
-    public int hashCode() {
-
-        return super.hashCode();
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-
-        return super.equals(obj);
-    }
-
-    @Override
-    protected Object clone() throws CloneNotSupportedException {
-
-        return super.clone();
-    }
-
-    @Override
-    public String toString() {
-
-        return super.toString();
-    }
-
-    @Override
-    public void visit(Token token) {
-
-        this.currentToken = token.text();
-
-    }
-
     private void setChildNode(Node node) {
 
         this.currentNode.setNextNode(node);
@@ -1631,9 +655,9 @@ public class ProgramAnalyzerNodeVisitor extends NodeVisitor {
         } else if (expressionNode.kind() == SyntaxKind.EXPLICIT_NEW_EXPRESSION) {
             ExplicitNewExpressionNode node = (ExplicitNewExpressionNode) expressionNode;
             if (!node.parenthesizedArgList().arguments().isEmpty()) {
-                visitSyntaxNode(node);
-                url = extractURL(this.currentToken);
-                this.currentToken = null;
+                UrlFinder urlFinder = new UrlFinder();
+                node.accept(urlFinder);
+                url = extractURL(urlFinder.getToken());
             }
         } else if (expressionNode.kind() == SyntaxKind.CHECK_EXPRESSION) {
             ExpressionNode expNode = ((CheckExpressionNode) expressionNode).expression();
@@ -1661,20 +685,23 @@ public class ProgramAnalyzerNodeVisitor extends NodeVisitor {
 
     private String findBaseUrl(SeparatedNodeList<FunctionArgumentNode> arguments) {
         // Assuming that the first argument is the base URL.
-        FunctionArgumentNode argumentNode = arguments.get(0);
+        for (FunctionArgumentNode argumentNode : arguments) {
+            UrlFinder urlFinder = new UrlFinder();
+            argumentNode.accept(urlFinder);
 
-        if (argumentNode.kind() == SyntaxKind.POSITIONAL_ARG &&
-                ((PositionalArgumentNode) argumentNode).expression().kind() == SyntaxKind.STRING_LITERAL) {
-            visitSyntaxNode(argumentNode);
-            String url = extractURL(this.currentToken);
-            this.currentToken = null;
-            return url;
-        } else if (argumentNode.kind() == SyntaxKind.NAMED_ARG &&
-                ((NamedArgumentNode) argumentNode).expression().kind() == SyntaxKind.STRING_LITERAL) {
-            visitSyntaxNode(argumentNode);
-            String url = extractURL(this.currentToken);
-            this.currentToken = null;
-            return url;
+            String url = urlFinder.getToken();
+            if (url.isEmpty()) {
+                continue;
+            }
+            if (argumentNode.kind() == SyntaxKind.POSITIONAL_ARG &&
+                    ((PositionalArgumentNode) argumentNode).expression().kind() == SyntaxKind.STRING_LITERAL) {
+                visitSyntaxNode(argumentNode);
+                return extractURL(url);
+            } else if (argumentNode.kind() == SyntaxKind.NAMED_ARG &&
+                    ((NamedArgumentNode) argumentNode).expression().kind() == SyntaxKind.STRING_LITERAL) {
+                visitSyntaxNode(argumentNode);
+                return extractURL(url);
+            }
         }
         return null;
     }
